+++ src/common/interfaces/nest-application.interface.ts
@@ -81,6 +81,13 @@ export interface INestApplication extends INestApplicationContext {
     getMicroservices(): INestMicroservice[];
 
     /**
+     * Returns underlying native HTTP server.
+     *
+     * @returns http.Server
+     */
+    getHttpServer(): any;
+
+    /**
      * Starts all the connected microservices asynchronously
      *
      * @param  {Function} callback Optional callback function